% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrsquared.R
\docType{package}
\name{mvrsquared}
\alias{mvrsquared}
\alias{mvrsquared-package}
\title{mvrsquared}
\description{
Compute the Coefficient of Determination for Vector or Matrix Outcomes
}
\details{
[DESCRIPTION HERE]
}
